sam_utils.py
from PIL import ImageDraw

# Function to segment leaf using SAM
def apply_sam_segmentation(image, pred_boxes, model):
    # Placeholder for SAM segmentation
    draw = ImageDraw.Draw(image)

    # Draw bounding boxes as rough segmentation
    for box in pred_boxes:
        x_min, y_min, x_max, y_max = box.tolist()
        draw.rectangle([x_min, y_min, x_max, y_max], outline="blue", width=3)

    return image

# Example loading function (Replace with actual SAM model)
def load_sam_model():
    # Replace with your model loading code
    return None
