requirements.txt
torch
transformers
numpy
matplotlib
Pillow
opencv-python
huggingface_hub
